refatorando para ficar mais top

+++ src/routes.js
@@ -10,10 +10,15 @@ export const routes = [
     method: 'POST',
     path: buildRoutePath('/tasks'),
     handler: (req, res) => {
-      const {
-        title,
-        description
-      } = req.body
+      const { title, description } = req.body
+
+      if (!title) {
+        return res.writeHead(400).end(JSON.stringify({ message: 'title is required' }))
+      }
+
+      if (!description) {
+        return res.writeHead(400).end(JSON.stringify({ message: 'description is required' }))
+      }
 
       const task = {
         id: randomUUID(),
@@ -33,7 +38,13 @@ export const routes = [
     method: 'GET',
     path: buildRoutePath('/tasks'),
     handler: (req, res) => {
-      const tasks = database.select('tasks')
+      const { search } = req.query
+
+      const tasks = database.select('tasks', {
+        title: search,
+        description: search,
+      })
+
       return res.end(JSON.stringify(tasks))
     }
   },
@@ -44,9 +55,22 @@ export const routes = [
       const { id } = req.params
       const { title, description } = req.body
 
+      if (!title && !description) {
+        return res.writeHead(400).end(
+          JSON.stringify({ message: 'title or description are required' })
+        )
+      }
+
+      const [ task ] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
+
       database.update('tasks', id, {
-        title,
-        description
+        title: title ?? task.title,
+        description: description ?? task.description,
+        updated_at: new Date(),
       })
 
       return res.writeHead(204).end()
@@ -58,7 +82,14 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params
 
+      const [ task ] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(400).end()
+      }
+
       database.delete('tasks', id)
+
       return res.writeHead(204).end()
     }
   },
@@ -68,7 +99,17 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params
 
-      database.complete('tasks', id)
+      const [ task ] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(400).end()
+      }
+
+      // implantando um toggle
+      const isTaskCompleted = !!task.completed_at
+      const completed_at = isTaskCompleted ? null : new Date();
+
+      database.update('tasks', id, { completed_at })
 
       return res.writeHead(204).end()
     }
